Set up my example hands-on project to use new Products entity and remove Custom PermissionEvaluator and other classes that are outside the scope of the next couple examples
In this simplified example, we store the owner as a string with a
constraint that it must be a username. I've learned through this setup
that if we want to, say, use a foreign key to the user ID (instead of
just thier username) and use that information in spring security, we
need to extend certain spring security classes/interfaces. Name the
User, UserDetails, UserDetailsService, or a combination there of and
maybe more. for now, we'll just use the username