@PreFilter. Custom Principal attributes, Fix CustomAuthProvider
Ok this commit has 3 things going on 1. We're learning to use the @PreFilter annotation to filter method arguments based on the security context 2. We've fix the CustomAuthProvider, before we mistakenly were just passing the username as the principal in the Authentication object instead of the user object. 3. Fixes the CustomAuthProvider allows us to add custom attributes to our principal so we can do more than just check against the current user's username